Fix AP Parameters Characteristic writing behaviour
The initial implementation was connecting to the Wi-Fi Access Point on
writing the AP Parameters.

That Characteristic is for saving the AP Parameters; the Connection
State Characteristic is for managing the connection.

This fixes the behaviour by connecting to the Wi-Fi Access Point on
writing CONNECT (0x01) to the Connection State Characteristic.